refactor(usage): extract usage payload builder into app/usage_service.py
